refactor: Add anotation @Entity
